+++ fili-core/src/main/java/com/yahoo/bard/webservice/web/MetricParser.java
@@ -2,14 +2,14 @@
 // Licensed under the terms of the Apache license. Please see LICENSE.md file distributed with this work for terms.
 package com.yahoo.bard.webservice.web;
 
-import org.json.JSONArray;
-import org.json.JSONException;
+import com.fasterxml.jackson.databind.node.ArrayNode;
+import com.fasterxml.jackson.databind.node.JsonNodeFactory;
+import com.fasterxml.jackson.databind.node.ObjectNode;
+
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import java.util.ArrayList;
 import java.util.HashMap;
-import java.util.List;
 import java.util.Map;
 import java.util.Stack;
 
@@ -21,7 +21,7 @@ public class MetricParser {
     private static final Logger LOG = LoggerFactory.getLogger(MetricParser.class);
 
     /**
-     * This function converts the string of metrics extracted from the url into JSONArray.
+     * This function converts the string of metrics extracted from the url into ArrayNode.
      *
      * @param metricString  An Api metric string eg:
      * <pre>{@code metricString = metric1(AND(dim1|id-in[a,b],dim2|id-in[c,d])),metric2 }</pre>
@@ -30,10 +30,9 @@ public class MetricParser {
      * <pre>{@code {[{"filter":{"AND":{"dim2|id-in":["abc","xyz"],"dim3|id-in":["mobile","tablet"]}}, "name":"metric1"},
      * {"filter":{},"name":"metric2"}]}}</pre>
 
-     * @throws JSONException if a JSON cannot be constructed successfully from the given metricString
      * @throws IllegalArgumentException if metricString is empty or the metricString has unbalanced brackets
      */
-    public static JSONArray generateMetricFilterJsonArray(String metricString) throws JSONException {
+    public static ArrayNode generateMetricFilterJsonArray(String metricString) {
         if (metricString.length() == 0 || !(isBracketsBalanced(metricString))) {
             LOG.error("Metrics parameter values are invalid. The string is: " + metricString);
             throw new IllegalArgumentException("Metrics parameter values are invalid. The string is: " + metricString);
@@ -45,7 +44,7 @@ public class MetricParser {
         String[] metrics = encodeMetricFilters(metricString, mapper).split(",");
 
         //looping each metric in a metrics array
-        List<String> metricList = new ArrayList<>();
+        ArrayNode arrayNode = JsonNodeFactory.instance.arrayNode();
         for (String metric : metrics) {
             //Separating metric and '-' separated filter encode metricFilterArray[0] will be metric name
             //and metricFilterArray[1] will be key of the mapper contains respective filter string of the metric
@@ -53,34 +52,32 @@ public class MetricParser {
             //Retrieving the filter string from the mapper and creating a metric set with metric as key
             //and filter as value
             String metricFilter = metricFilterArray.length > 1 ? mapper.get("-" + metricFilterArray[1]) : "";
-            metricList.add(getJsonString(metricFilterArray[0], metricFilter));
+
+            ObjectNode objectNode = arrayNode.objectNode();
+            addJsonFilter(objectNode, metricFilterArray[0], metricFilter);
+            arrayNode.add(objectNode);
         }
-        StringBuilder finalJson = new StringBuilder();
-        //preparing complete json string by appending brackets to metric name and metric filter
-        for (String jsonToken : metricList) {
-            finalJson.append("{" + jsonToken + "},");
-        }
-        return new JSONArray("[" + finalJson.toString().substring(0, finalJson.toString().length() - 1) + "]");
+
+        return arrayNode;
     }
 
     /**
-     * Returns a String that is formatted to a JSON string.
+     * Add the name and filter of a metric to an ObjectNode.
+     * For eg: {"filter":{"AND":"property|id-in:[14,125],country|id-in:[US,IN]},"name":"foo"}
      *
+     * @param metricNode  the object node to fill with the name and filter on a metric
      * @param metricName  the name of the metric
      * @param metricFilter  the filter associated with the metric. Could be empty or have a value
-     *
-     * @return A String in JSON format for a given metric and filter.
-     * For eg: "filter":{"AND":"property|id-in:[14,125],country|id-in:[US,IN]},"name":"foo"
      */
-    private static String getJsonString(String metricName, String metricFilter) {
-        if (metricFilter.isEmpty()) {
-            metricFilter = "\"filter\": {}";
-        } else {
-            metricFilter = "\"filter\":{\"AND\":\"" +
-                    metricFilter.replace("AND", "").replace("(", "").replace(")", "") + "\" }";
-
+    private static void addJsonFilter(ObjectNode metricNode, String metricName, String metricFilter) {
+        metricNode.put("name", metricName);
+        ObjectNode filterNode = metricNode.putObject("filter");
+        if (!metricFilter.isEmpty()) {
+            metricFilter = metricFilter.replace("AND", "")
+                    .replace("(", "")
+                    .replace(")", "");
+            filterNode.put("AND", metricFilter);
         }
-        return "\"name\": \"" + metricName + "\"," + metricFilter;
     }
 
     /**